requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
einops==0.8.1
filelock==3.18.0
filetype==1.2.0
fsspec==2025.5.1
ftfy==6.3.1
google-auth==2.40.3
google-genai==1.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
identify==2.6.12
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
markdown2==2.5.3
markdownify==1.1.0
marker-pdf==1.8.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
openai==1.93.0
opencv-python-headless==4.11.0.86
packaging==25.0
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.3.8
pre_commit==4.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdfium2==4.30.0
python-dotenv==1.1.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
surya-ocr==0.14.6
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
